requirements.txt
altair==6.1.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.6
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.47
idna==3.12
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.20.0
numpy==2.4.4
packaging==26.1
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
pyarrow==24.0.0
pydeck==0.9.2
python-dateutil==2.9.0.post0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
streamlit-drawable-canvas==0.9.3
stripe==15.0.1
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
watchdog==6.0.0
